abstract message list scroll to hook
