--- apps/client/src/components/games/Uno/components/UnoGame/components/TurnIndicator/TurnIndicator.module.css
@@ -1,65 +1,0 @@
-.container {
-	--initial-transform: translateX(-50%);
-	--bg-color: var(--uno-color-blue);
-	--text-shadow-1: hsl(204, 85%, 15%);
-	--text-shadow-2: hsl(207, 85%, 12%);
-
-	--corner-clip: 1rem;
-
-	position: absolute;
-	left: 50%;
-	transform: var(--initial-transform);
-	top: min(15rem, 15%);
-	background-color: var(--bg-color);
-	color: var(--primary-text-white);
-	padding: 0.5rem 2rem;
-	font-size: 2rem;
-	z-index: 3;
-	animation: fadeIn 0.4s ease-in-out;
-	text-align: center;
-
-	clip-path: polygon(
-		0 var(--corner-clip),
-		var(--corner-clip) 0,
-		calc(100% - var(--corner-clip)) 0,
-		100% var(--corner-clip),
-		100% calc(100% - var(--corner-clip)),
-		calc(100% - var(--corner-clip)) 100%,
-		var(--corner-clip) 100%,
-		0% calc(100% - var(--corner-clip)),
-		0% var(--corner-clip)
-	);
-
-	&.ourTurn {
-		--bg-color: var(--uno-color-green);
-		--text-shadow-1: hsl(104, 85%, 15%);
-		--text-shadow-2: hsl(107, 85%, 12%);
-	}
-}
-
-@keyframes bounce {
-	0%,
-	20%,
-	50%,
-	80%,
-	100% {
-		transform: translateY(0);
-	}
-	40% {
-		transform: translateY(-30%);
-	}
-	60% {
-		transform: translateY(-15%);
-	}
-}
-
-@keyframes fadeIn {
-	from {
-		transform: var(--initial-transform) translateY(-100%) scale(0.8);
-		opacity: 0;
-	}
-	to {
-		transform: var(--initial-transform) translateY(0) scale(1);
-		opacity: 1;
-	}
-}
