+++ apps/client/src/pages/room/[roomCode]/components/Room/components/Sidebar/components/MessagesList/hooks/index.ts
@@ -1,0 +1,3 @@
+import useScrollToBottom from './useScrollToBottom';
+
+export { useScrollToBottom };
