+++ apps/client/src/pages/room/[roomCode]/components/Room/components/Sidebar/components/MessagesList/index.tsx
@@ -3,33 +3,16 @@ import { useRef, useContext, useEffect, useCallback } from 'react';
 import { AudioContext } from '@/components';
 import classes from './MessagesList.module.css';
 import MessageComponent from './Message';
+import { useScrollToBottom } from './hooks';
 
 type MessagesListProps = {
 	messages: Message[];
 	ourPlayer: Player;
 };
 const MessagesList = ({ messages, ourPlayer }: MessagesListProps) => {
-	const listRef = useRef<HTMLOListElement>(null);
+	const listRef = useScrollToBottom(messages);
 	const audioContext = useContext(AudioContext);
 	const scrollPositionRef = useRef<number>(0);
-
-	// Scroll to the bottom of the chat if user was at bottom before new message
-	const checkShouldScrollToBottom = (addedElement: HTMLLIElement) => {
-		const list = listRef.current as HTMLOListElement;
-
-		const lastMessageStyle = window.getComputedStyle(addedElement);
-		const marginTop = parseInt(lastMessageStyle.marginTop, 10);
-		const marginBottom = parseInt(lastMessageStyle.marginBottom, 10);
-
-		const lastMessageHeight = addedElement.offsetHeight + marginTop + marginBottom;
-		const messagesListHeight = list.scrollHeight;
-		const heightBeforeLastMessage = messagesListHeight - lastMessageHeight;
-		const scrollPosition = list.scrollTop + list.clientHeight;
-
-		if (scrollPosition >= heightBeforeLastMessage - 10) {
-			list.scrollTo({ top: messagesListHeight, behavior: 'auto' });
-		}
-	};
 
 	const checkShouldPlayAudioNotification = useCallback((message: Message) => {
 		if (message.type === 'user' && message.socketId !== ourPlayer.socketId) {
@@ -40,11 +23,6 @@ const MessagesList = ({ messages, ourPlayer }: MessagesListProps) => {
 	}, [audioContext, ourPlayer.socketId]);
 
 	useEffect(() => {
-		const lastMessageEl = listRef.current?.lastElementChild as HTMLLIElement | null;
-		if (lastMessageEl) {
-			checkShouldScrollToBottom(lastMessageEl);
-		}
-
 		const lastMessage = messages[messages.length - 1];
 		if (lastMessage) {
 			checkShouldPlayAudioNotification(lastMessage);
